Fix bug where validity check fails for certain vector sizes.
This happened in the case where the length of the input chunk
to the parallel sum gadget was not of the required length, and
so had to be extended. What was wrong is that the extension
happened with the field element zero, whereas it should have
happened with (a share) of the encoded zero value, which is
`2^(n-1)`. This is now fixed.

Also added a test which tests the `fixedpoint_l2` type on long
vectors, as this is how this bug was discovered.

Also moved the testing function `test_fixed` into the outer scope.